Add name to union_column to ease override
The current union_column uses the field.name to get the union_field, it works
correctly but prevent to implement by override a clean mapping transformation of
name between the union and the models. Indeed it is doable by passing an other
field but then we could potentially lost the automatic sharding.

issue3992
review13331003