Use NOWAIT when locking a table
lock is used to enforce a serializable level for all transaction using on the
same table. This is not achieved if transactions are waiting for the lock
instead of abort and retry with a new transaction because they will still have
the initial transaction row values with the repeatable read level.

issue4373
review5901002
(grafted from eba810510a4a5ae8ed20668733c1af5750cd6788)